+++ src/SMAPI.Web/Framework/Metrics/ApiMetricsModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using StardewModdingAPI.Toolkit.Framework.UpdateData;
 
@@ -15,14 +16,28 @@ namespace StardewModdingAPI.Web.Framework.Metrics
         /// <summary>The metrics by mod site.</summary>
         public Dictionary<ModSiteKey, MetricsModel> Sites { get; } = new();
 
+        /// <summary>The number of update-check requests by SMAPI version.</summary>
+        public Dictionary<string, long> ByApiVersion { get; } = new(StringComparer.OrdinalIgnoreCase);
+
+        /// <summary>The number of update-check requests by game version.</summary>
+        public Dictionary<string, long> ByGameVersion { get; } = new(StringComparer.OrdinalIgnoreCase);
+
 
         /*********
         ** Public methods
         *********/
         /// <summary>Track an update-check request received by the API.</summary>
-        public void TrackRequest()
+        /// <param name="apiVersion">The SMAPI version installed by the player.</param>
+        /// <param name="gameVersion">The game version installed by the player.</param>
+        public void TrackRequest(ISemanticVersion? apiVersion, ISemanticVersion? gameVersion)
         {
             this.ApiRequests++;
+
+            string apiVersionStr = apiVersion?.ToString() ?? "<none specified>";
+            string gameVersionStr = gameVersion?.ToString() ?? "<none specified>";
+
+            this.ByApiVersion[apiVersionStr] = this.ByApiVersion.GetValueOrDefault(apiVersionStr) + 1;
+            this.ByGameVersion[gameVersionStr] = this.ByGameVersion.GetValueOrDefault(gameVersionStr) + 1;
         }
 
         /// <summary>Track the update-check result for a specific update key.</summary>