+++ src/SMAPI.Web/Framework/Metrics/MetricsSummary.cs
@@ -13,6 +13,8 @@ namespace StardewModdingAPI.Web.Framework.Metrics
     /// <param name="TotalCacheHits">The number of times an update key returned data from the cache.</param>
     /// <param name="TotalSuccessCacheMisses">The number of times an update key successfully fetched data from a remote mod site.</param>
     /// <param name="TotalErrorCacheMisses">The number of times an update key could not fetch data from a remote mod site (e.g. mod page didn't exist or mod site returned an API error).</param>
+    /// <param name="ByApiVersion">The number of update-check request by SMAPI version.</param>
+    /// <param name="ByGameVersion">The number of update-check request by game version.</param>
     /// <param name="BySite">The metrics grouped by site.</param>
     /// <param name="ByDate">The metrics grouped by UTC date.</param>
     internal record MetricsSummary(
@@ -24,6 +26,8 @@ namespace StardewModdingAPI.Web.Framework.Metrics
         int TotalCacheHits,
         int TotalSuccessCacheMisses,
         int TotalErrorCacheMisses,
+        IDictionary<string, long> ByApiVersion,
+        IDictionary<string, long> ByGameVersion,
         IDictionary<ModSiteKey, MetricsModel> BySite,
         IDictionary<string, ApiMetricsModel> ByDate
     );